Moved interfaces to contracts folder
